+++ R/zzz.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("part", "group", "cluster"))
+utils::globalVariables(c("part", "group", "cluster", "label", "isLargestPart"))